+++ pylabrobot/centrifuge/__init__.py
@@ -14,5 +14,4 @@ from .standard import (
   LoaderNoPlateError,
   NotAtBucketError,
 )
-from .v11_vspin_backend import V11VSpinBackend
 from .vspin_backend import Access2Backend, VSpinBackend, create_vspin_backend
